requirements.txt
pinecone-client
langchain
langchain-community
langchain-pinecone
langchain-huggingface
sentence-transformers
python-dotenv
streamlit
chainlit
